general refactorings around orgs and products and some cleanup
