+++ requisite-api/src/resources/organizations/memberships/OrgMembershipsCreateResource.ts
@@ -1,12 +1,12 @@
-import { getLogger } from '../../util/Logger';
-import ResourceRequest from '../../common/ResourceRequest';
+import { getLogger } from '../../../util/Logger';
+import ResourceRequest from '../../../common/ResourceRequest';
 import { Response, NextFunction } from 'express';
-import ServiceProvider from '../../services/ServiceProvider';
+import ServiceProvider from '../../../services/ServiceProvider';
 import Organization from '@requisite/model/lib/org/Organization';
 import { assertExists } from '@requisite/utils/lib/validation/AssertionUtils';
 import Membership from '@requisite/model/lib/user/Membership';
 
-const logger = getLogger('resources/organization-memberships/OrgMembershipsCreateResource');
+const logger = getLogger('resources/organizations/memberships/OrgMembershipsCreateResource');
 
 export default (req: ResourceRequest, res: Response, next: NextFunction): void => {
     (async function() {
