+++ requisite-api/src/resources/organizations/memberships/OrgMembershipsDeleteResource.ts
@@ -1,12 +1,12 @@
-import { getLogger } from '../../util/Logger';
-import ResourceRequest from '../../common/ResourceRequest';
+import { getLogger } from '../../../util/Logger';
+import ResourceRequest from '../../../common/ResourceRequest';
 import { Response, NextFunction } from 'express';
-import ServiceProvider from '../../services/ServiceProvider';
+import ServiceProvider from '../../../services/ServiceProvider';
 import { assertExists } from '@requisite/utils/lib/validation/AssertionUtils';
 import Organization from '@requisite/model/lib/org/Organization';
 import Membership from '@requisite/model/lib/user/Membership';
 
-const logger = getLogger('resources/organization-memberships/OrgMembershipsDeleteResource');
+const logger = getLogger('resources/organizations/memberships/OrgMembershipsDeleteResource');
 
 export default (req: ResourceRequest, res: Response, next: NextFunction): void => {
     (async function() {
