+++ requisite-api/src/resources/organizations/memberships/OrgMembershipsUpdateResource.ts
@@ -1,30 +1,30 @@
-import { getLogger } from '../../util/Logger';
-import ResourceRequest from '../../common/ResourceRequest';
+import { getLogger } from '../../../util/Logger';
+import ResourceRequest from '../../../common/ResourceRequest';
 import { Response, NextFunction } from 'express';
-import ServiceProvider from '../../services/ServiceProvider';
+import ServiceProvider from '../../../services/ServiceProvider';
+import Organization from '@requisite/model/lib/org/Organization';
 import { assertExists } from '@requisite/utils/lib/validation/AssertionUtils';
 import Membership from '@requisite/model/lib/user/Membership';
-import { ConflictError } from '../../util/ApiErrors';
-import Product from '@requisite/model/lib/product/Product';
+import { ConflictError } from '../../../util/ApiErrors';
 
-const logger = getLogger('resources/product-memberships/ProdMembershipsUpdateResource');
+const logger = getLogger('resources/organizations/memberships/OrgMembershipsUpdateResource');
 
 export default (req: ResourceRequest, res: Response, next: NextFunction): void => {
     (async function() {
         try {
-            logger.debug('Executing product memberships update resource');
+            logger.debug('Executing org memberships create resource');
             assertExists(req.entity, 'req.entity');
-            const membership = req.entity as Membership<Product>;
+            const membership = req.entity as Membership<Organization>;
 
-            const modMembership: Membership<Product>
-                = req.body as Membership<Product>;
+            const modMembership: Membership<Organization>
+                = req.body as Membership<Organization>;
 
             const membershipIdConflict =
                 modMembership.id !== null
                 && modMembership.id !== undefined
                 && modMembership.id !== membership.id;
 
-            const membershipEntityIdConfict =
+            const entityIdConfict =
                 modMembership.entity
                 && modMembership.entity.id !== null
                 && modMembership.entity.id !== undefined
@@ -38,7 +38,7 @@ export default (req: ResourceRequest, res: Response, next: NextFunction): void =
 
             if (membershipIdConflict) {
                 next(new ConflictError('The membership identifier in the body does not match the uri.'));
-            } else if (membershipEntityIdConfict) {
+            } else if (entityIdConfict) {
                 next(new ConflictError('The entity identifier in the body does not match the entity on the membership from the uri.  Entities cannot be changed on a membership'));
             } else if (userIdConfict) {
                 next(new ConflictError('The user identifier in the body does not match the user on the membership from the uri.  Users cannot be changed on a membership'));
@@ -46,7 +46,7 @@ export default (req: ResourceRequest, res: Response, next: NextFunction): void =
                 modMembership.id = membership.id;
                 modMembership.entity = membership.entity;
                 modMembership.user = membership.user;
-                await ServiceProvider.getProductsService()
+                await ServiceProvider.getOrganizationsService()
                     .updateMembership(modMembership);
                 res.status(200).send(modMembership);
             }
