+++ requisite-api/tests/orgs/memberships/supertest.app.orgs.memberships.item.get.test.ts
@@ -1,12 +1,12 @@
-import '../supertest.mock.sqlz';
-import '../supertest.mock.jsonwebtoken';
+import '../../supertest.mock.sqlz';
+import '../../supertest.mock.jsonwebtoken';
 import request from 'supertest';
-import { getApp } from '../../src/app';
-import { configure } from '../../src/util/Logger';
+import { getApp } from '../../../src/app';
+import { configure } from '../../../src/util/Logger';
 import { ValidationResult } from '@requisite/utils/lib/validation/ValidationUtils';
 import Organization from '@requisite/model/lib/org/Organization';
 import Membership from '@requisite/model/lib/user/Membership';
-import { getMockedOrg, getMockedOrgMembership, getMockedAuthBearerForUser, getMockedAuthBearerForOrgMembership, getMockedAuthBearerSystemAdmin } from '../mockUtils';
+import { getMockedOrg, getMockedOrgMembership, getMockedAuthBearerForUser, getMockedAuthBearerForOrgMembership, getMockedAuthBearerSystemAdmin } from '../../mockUtils';
 
 configure('ERROR');
