+++ requisite-api/tests/orgs/memberships/supertest.app.orgs.memberships.list.post.test.ts
@@ -1,10 +1,10 @@
-import '../supertest.mock.sqlz';
-import '../supertest.mock.jsonwebtoken';
+import '../../supertest.mock.sqlz';
+import '../../supertest.mock.jsonwebtoken';
 import request from 'supertest';
-import { getApp } from '../../src/app';
-import { configure } from '../../src/util/Logger';
+import { getApp } from '../../../src/app';
+import { configure } from '../../../src/util/Logger';
 import { ValidationResult } from '@requisite/utils/lib/validation/ValidationUtils';
-import { getAuthBearer, getMockedAuthBearerForOrgMembership, getMockedAuthBearerForUser, getMockedAuthBearerSystemAdmin, getMockedOrg, getMockedOrgMemberships, getMockedUser, getMockedUserForOrgMembership } from '../mockUtils';
+import { getAuthBearer, getMockedAuthBearerForOrgMembership, getMockedAuthBearerForUser, getMockedAuthBearerSystemAdmin, getMockedOrg, getMockedOrgMemberships, getMockedUser, getMockedUserForOrgMembership } from '../../mockUtils';
 import Membership, { OrganizationRole } from '@requisite/model/lib/user/Membership';
 import Organization from '@requisite/model/lib/org/Organization';
