+++ requisite-api/tests/products/memberships/supertest.app.products.memberships.item.get.test.ts
@@ -1,12 +1,12 @@
-import '../supertest.mock.sqlz';
-import '../supertest.mock.jsonwebtoken';
+import '../../supertest.mock.sqlz';
+import '../../supertest.mock.jsonwebtoken';
 import request from 'supertest';
-import { getApp } from '../../src/app';
-import { configure } from '../../src/util/Logger';
+import { getApp } from '../../../src/app';
+import { configure } from '../../../src/util/Logger';
 import { ValidationResult } from '@requisite/utils/lib/validation/ValidationUtils';
 import Organization from '@requisite/model/lib/org/Organization';
 import Membership, { ProductRole } from '@requisite/model/lib/user/Membership';
-import { getMockedProductMembership, getMockedProduct, getMockedAuthBearerForProductMembership, getMockedAuthBearerSystemAdmin, getMockedAuthBearerForUser, getMockedAuthBearerForOrgMembership } from '../mockUtils';
+import { getMockedProductMembership, getMockedProduct, getMockedAuthBearerForProductMembership, getMockedAuthBearerSystemAdmin, getMockedAuthBearerForUser, getMockedAuthBearerForOrgMembership } from '../../mockUtils';
 import Product from '@requisite/model/lib/product/Product';
 
 configure('ERROR');
