+++ requisite-api/tests/products/memberships/supertest.app.products.memberships.list.post.test.ts
@@ -1,11 +1,11 @@
-import '../supertest.mock.sqlz';
-import '../supertest.mock.jsonwebtoken';
+import '../../supertest.mock.sqlz';
+import '../../supertest.mock.jsonwebtoken';
 import request from 'supertest';
-import { getApp } from '../../src/app';
-import { configure } from '../../src/util/Logger';
+import { getApp } from '../../../src/app';
+import { configure } from '../../../src/util/Logger';
 import { ValidationResult } from '@requisite/utils/lib/validation/ValidationUtils';
 import Product from '@requisite/model/lib/product/Product';
-import { getMockedAuthBearerForProductMembership, getMockedAuthBearerForUser, getMockedAuthBearerSystemAdmin, getMockedProduct, getMockedProductMemberships, getMockedUserForProductMembership } from '../mockUtils';
+import { getMockedAuthBearerForProductMembership, getMockedAuthBearerForUser, getMockedAuthBearerSystemAdmin, getMockedProduct, getMockedProductMemberships, getMockedUserForProductMembership } from '../../mockUtils';
 import Organization from '@requisite/model/lib/org/Organization';
 import Membership, { ProductRole } from '@requisite/model/lib/user/Membership';
